requirements.txt
python-dotenv>=1.0.1
openai>=1.75.0
langchain>=0.3.25
langchain-openai>=0.3.14
pypdf>=5.4.0
python-docx>=1.1.2
pandas>=2.2.3
openpyxl>=3.1.5
